Moved and readapted tests
