+++ remoteappmanager/rest/tests/test_registry.py
