+++ remoteappmanager/tests/fixtures/templates/foo.html
