+++ remoteappmanager/tests/mocking/virtual/docker_client.py
