--- tests/handlers/test_home_handler.py
@@ -1,52 +1,0 @@
-import os
-
-from tests import utils
-from tests.mocking import dummy
-from tests.temp_mixin import TempMixin
-
-
-class TestHomeHandler(TempMixin, utils.AsyncHTTPTestCase):
-    def setUp(self):
-        self._old_proxy_api_token = os.environ.get("PROXY_API_TOKEN", None)
-        os.environ["PROXY_API_TOKEN"] = "dummy_token"
-
-        def cleanup():
-            if self._old_proxy_api_token is not None:
-                os.environ["PROXY_API_TOKEN"] = self._old_proxy_api_token
-            else:
-                del os.environ["PROXY_API_TOKEN"]
-
-        self.addCleanup(cleanup)
-
-        super().setUp()
-
-    def get_app(self):
-        app = dummy.create_application()
-        app.hub.verify_token.return_value = {
-            'pending': None,
-            'name': app.settings['user'],
-            'admin': False,
-            'server': app.settings['base_urlpath']}
-        return app
-
-    def test_home(self):
-        res = self.fetch("/user/username/",
-                         headers={
-                             "Cookie": "jupyter-hub-token-username=foo"
-                         }
-                         )
-
-        self.assertEqual(res.code, 200)
-        self.assertIn("Available Applications", str(res.body))
-
-    def test_failed_auth(self):
-        self._app.hub.verify_token.return_value = {}
-        res = self.fetch("/user/username/",
-                         headers={
-                             "Cookie": "jupyter-hub-token-username=foo"
-                         }
-                         )
-
-        self.assertGreaterEqual(res.code, 400)
-        self.assertIn(self._app.file_config.login_url, res.effective_url)
-        self.assertNotIn("Available Applications", str(res.body))
